// src/api/models/pet-im.ts
/* tslint:disable */
/* eslint-disable */
/**
 * APF.API
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: 1.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
/**
 * 
 * @export
 * @interface PetIM
 */
export interface PetIM {
    /**
     * 
     * @type {string}
     * @memberof PetIM
     */
    animalName: string;
    /**
     * 
     * @type {number}
     * @memberof PetIM
     */
    age: number;
    /**
     * 
     * @type {number}
     * @memberof PetIM
     */
    weight: number;
    /**
     * 
     * @type {string}
     * @memberof PetIM
     */
    timer: string;
    /**
     * 
     * @type {string}
     * @memberof PetIM
     */
    day: string;
}