// src/api/models/pet-vm.ts
/* tslint:disable */
/* eslint-disable */
/**
 * APF.API
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: 1.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
/**
 * 
 * @export
 * @interface PetVM
 */
export interface PetVM {
    /**
     * 
     * @type {string}
     * @memberof PetVM
     */
    id?: string | null;
    /**
     * 
     * @type {string}
     * @memberof PetVM
     */
    animalName?: string | null;
    /**
     * 
     * @type {number}
     * @memberof PetVM
     */
    age?: number;
    /**
     * 
     * @type {number}
     * @memberof PetVM
     */
    weight?: number;
    /**
     * 
     * @type {string}
     * @memberof PetVM
     */
    timer?: string | null;
    /**
     * 
     * @type {string}
     * @memberof PetVM
     */
    day?: string | null;
    /**
     * 
     * @type {boolean}
     * @memberof PetVM
     */
    isDeleted?: boolean;
    /**
     * 
     * @type {boolean}
     * @memberof PetVM
     */
    isActive?: boolean;
}